Start of strictly typed implementation
