--- experiments/metrics.py
@@ -1,5 +1,0 @@
-""" File for units and conversion """
-
-from pint import UnitRegistry
-units = UnitRegistry()
-quantity = units.Quantity
