+++ experiments/mypy.ini
@@ -1,0 +1,18 @@
+# Global options:
+[mypy]
+python_version=3.9
+
+[mypy-isotonic.isotonic.*]
+ignore_missing_imports = True
+[mypy-progressbar.*]
+ignore_missing_imports = True
+[mypy-yappi.*]
+ignore_missing_imports = True
+[mypy-scipy.interpolate.*]
+ignore_missing_imports = True
+[mypy-sklearn.isotonic.*]
+ignore_missing_imports = True
+[mypy-sklearn.metrics.*]
+ignore_missing_imports = True
+[mypy-scipy.signal.*]
+ignore_missing_imports = True
